traffic-recognition-v2/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6
opencv-python-headless>=4.8.0.74
numpy>=1.24.0
pillow>=10.0.0 